Update and rename signUpTest.java to SignupValidatorTest.java
added package, resolved error
